remove processing queue, merge ParticleProcessor logic into fluence client
