+++ src/api.unstable.ts
@@ -82,6 +82,7 @@ export const createClient = async (
     }
 
     const client = new ClientImpl(peerId);
+    await client.initAquamarineRuntime();
 
     if (connectTo) {
         let theAddress: Multiaddr;
@@ -96,8 +97,6 @@ export const createClient = async (
         if (!(await checkConnection(client))) {
             throw new Error('Connection check failed. Check if the node is working or try to connect to another node');
         }
-    } else {
-        await client.local();
     }
 
     return client;
